# Make slice decoding more performant 💨