This PR makes a few tweaks to improve the performance of decoding into a slice
- When decoding a slice, a bug meant we were doing double allocations of slice values and then throwing the second set away (because they were never used)
- When we initialize/create a struct element, it comes pre-allocated with the zero values of elements. We can use those allocations rather than allocate ourselves again. This also saves some double allocations and gives as a bonus of not having to re-iterate and set the pointers (so instead of an intermediate variable to decode from struct element to intermediate variable to struct element), we just pass the direct struct element to be decoded into

This code is already [well covered with tests](https://github.com/monzo/gocassa/blob/master/scanner_test.go) so that gives confidence that nothing has changed behaviour wise.

Before:
```
pkg: github.com/monzo/gocassa
BenchmarkDecodeBlogSliceNoBody
BenchmarkDecodeBlogSliceNoBody-8       	   30699	     38875 ns/op	   26376 B/op	     388 allocs/op
BenchmarkDecodeBlogStruct
BenchmarkDecodeBlogStruct-8            	  364782	      3258 ns/op	    2992 B/op	      33 allocs/op
BenchmarkDecodeBlogStructEmptyBody
BenchmarkDecodeBlogStructEmptyBody-8   	  344191	      3236 ns/op	    2992 B/op	      33 allocs/op
BenchmarkDecodeAlphaSlice
BenchmarkDecodeAlphaSlice-8            	   10000	    118773 ns/op	   56507 B/op	    1324 allocs/op
BenchmarkDecodeAlphaStruct
BenchmarkDecodeAlphaStruct-8           	  126303	      9531 ns/op	    9226 B/op	      93 allocs/op
BenchmarkStatementMapTable
BenchmarkStatementMapTable-8           	  465386	      2499 ns/op	    3664 B/op	      33 allocs/op
```

After:
```
pkg: github.com/monzo/gocassa
BenchmarkDecodeBlogSliceNoBody
BenchmarkDecodeBlogSliceNoBody-8       	   42524	     29752 ns/op	   17232 B/op	     108 allocs/op
BenchmarkDecodeBlogStruct
BenchmarkDecodeBlogStruct-8            	  695108	      1559 ns/op	     736 B/op	       7 allocs/op
BenchmarkDecodeBlogStructEmptyBody
BenchmarkDecodeBlogStructEmptyBody-8   	  773036	      1586 ns/op	     736 B/op	       7 allocs/op
BenchmarkDecodeAlphaSlice
BenchmarkDecodeAlphaSlice-8            	   13658	     86213 ns/op	   30640 B/op	     208 allocs/op
BenchmarkDecodeAlphaStruct
BenchmarkDecodeAlphaStruct-8           	  263876	      4346 ns/op	    1132 B/op	      11 allocs/op
BenchmarkStatementMapTable
BenchmarkStatementMapTable-8           	  471937	      2551 ns/op	    3664 B/op	      33 allocs/op
```