@suhailpatel 2021-03-24
Whilst i've not added a substantial amount of tests in this particular PR, we have a [whole army of test cases](https://github.com/monzo/gocassa/blob/master/scanner_test.go) in the Scanner already testing the relevant refactored portions.